refactor: move  client setup to its own package
this moves the git provider client setup logic to
its own package gitclient so that we can call it
from github package. at the moment if we call client
setup from github it creates import cycle.
