+++ pkg/secrets/secret.go
@@ -1,4 +1,4 @@
-package pipelineascode
+package secrets
 
 import (
 	"context"
